+++ src/main/java/business/entity/Episode.java
@@ -3,19 +3,19 @@ package business.entity;
 public class Episode extends Video {
 
     private String title;    
-    private double time;
+    private int time;
 
-    public Episode(String title, double time) {
+    public Episode(String title, int time) {
         super(title);
         this.title = title;
         this.time = time;
     }
 
-    public double getTime() {
+    public int getTime() {
         return time;
     }
 
-    public void setTime(double time) {
+    public void setTime(int time) {
         this.time = time;
     }
