+++ src/main/java/data/IRepositoryVideo.java
@@ -9,7 +9,7 @@ import business.interfaces.GetVideo;
 
 public interface IRepositoryVideo extends GetVideo {
 
-    boolean createMovie(String name, double temp, String genre,int year, String nameCast);
+    boolean createMovie(String name, int temp, String genre,int year, String nameCast);
 
     boolean createSerie(String name, String genre,int year, String nameCast);
